fix for ShardedDataIterator's apply function
This function works incorrectly in the distributed training mod. It applies the visitor function only to a specific to a current processing node chunk of data and since all data are shuffled before every epoch, this lead to some data not being visitied.
Temporal fix before a more efficient solution.